backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
Flask-JWT-Extended==4.5.2
Flask-SQLAlchemy==3.0.5
PyMySQL==1.1.0
Werkzeug==2.3.7
python-dotenv==1.0.0
